CIT-24 enregistrer une vente

+++ src/main/java/com/wora/citronix/repositories/RecolteRepository.java
@@ -4,6 +4,6 @@ import com.wora.citronix.entities.Recolte;
 import com.wora.citronix.entities.enums.Saison;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecolteRepository extends JpaRepository<Recolte, Integer> {
+public interface RecolteRepository extends JpaRepository<Recolte, Long> {
     boolean existsBySaison(Saison season);
 }
